Finish unit-tests
We now have working unit-tests covering all the functions running
sparql queries. This allows us to tests the sparql queries and
ensure its output remains consistent.

The tests are dependant on the data returned so a single change
in the data returned by sparql would likely make the tests fail.